🚚 reorganized Footer Component into a single directory
