+++ libs/shared/data/src/lib/constants/author.constants.ts
@@ -3,8 +3,8 @@
  */
 export enum Author {
   FullName = 'Usama Ansari',
-  GitHub = 'usama251993',
-  LinkedIn = 'usama-ansari-81380080',
+  GitHub = 'https://github.com/usama251993',
+  LinkedIn = 'https://www.linkedin.com/in/usama-ansari-81380080',
   Discord = 'usama251993#5438',
   Email = 'usama251993@gmail.com',
   Location = 'https://goo.gl/maps/X6syXXRrRML4FUgc6'
