+++ libs/ui/layout/src/lib/components/footer/component/footer-container/footer-container.component.spec.ts
