+++ libs/ui/layout/src/lib/components/footer/imports.ts
@@ -1,0 +1,4 @@
+export type { IconModel, ImageModel } from '../../imports/models';
+export { IconStub } from '../../imports/constants';
+
+export { ClipboardService, SnackbarService } from '../../imports/services';
