+++ libs/ui/layout/src/lib/components/footer/index.ts
@@ -1,0 +1,20 @@
+export {
+  FooterAssetsStub,
+  FooterDataStub,
+  FooterAssets,
+  FooterStrings
+} from './constant/footer.constants';
+
+export type {
+  FooterAssetsModel,
+  FooterDataModel,
+  FooterIconType,
+  FooterLinkModel,
+  FooterSnackbarMessageModel,
+  FooterStringsModel
+} from './model/footer.model';
+
+export { FooterContainerComponent } from './component/footer-container/footer-container.component';
+export { FooterComponent } from './component/footer-container/footer/footer.component';
+
+export { FooterService } from './service/footer.service';
