+++ libs/ui/layout/src/lib/services/index.ts
@@ -1,2 +1,1 @@
-export { FooterService } from './footer/footer.service';
 export { ToolbarService } from './toolbar/toolbar.service';
